Added stdlib.jar
Using the standard libraries. The file stdlib.jar bundles together all of our standard libraries into one file. There are a number of ways to access the libraries:
Current directory. The easiest (but not the sanest) way to use the standard libraries to download stdlib.jar and unjar it in your current working directory.
% jar xf stdlib.jar
Alternatively, you can download the individual .java files you need (such as StdIn.java) and put them in the same directory as the program you are writing. Then, compile and execute as usual.
% javac MyProgram.java
% java  MyProgram
This approach has the drawback that you need a copy of each .java file you need in each directory where you need it.
Classpath. Put stdlib.jar in the same directory as the program you are writing (but do not unjar it). Then, compile and execute as follows:
OS X / Linux
------------
% javac -cp .:stdlib.jar MyProgram.java
% java  -cp .:stdlib.jar MyProgram

Windows
------------
% javac -cp .;stdlib.jar MyProgram.ja